+++ src/collective/documentgenerator/browser/controlpanel.py
@@ -7,10 +7,13 @@ from collective.documentgenerator.config import DEFAULT_OO_PORT
 from collective.documentgenerator.config import DEFAULT_OO_SERVER
 from collective.documentgenerator.config import DEFAULT_PYTHON_UNO
 from collective.documentgenerator.interfaces import IDocumentGeneratorSettings
+from collective.z3cform.datagridfield import DataGridFieldFactory
+from collective.z3cform.datagridfield.registry import DictRow
 from plone.app.registry.browser.controlpanel import ControlPanelFormWrapper
 from plone.app.registry.browser.controlpanel import RegistryEditForm
 from Products.CMFPlone.utils import safe_unicode
 from Products.statusmessages.interfaces import IStatusMessage
+from plone.autoform import directives
 from z3c.form import button
 from zope import schema
 from zope.interface import implementer
@@ -45,6 +48,14 @@ def check_for_uno(value):
     return True
 
 
+class IOOPortRowSchema(Interface):
+    oo_port = schema.Int(
+        title=_(u'oo_port'),
+        description=_(u'Port Number of OO.'),
+        required=True,
+    )
+
+
 class IDocumentGeneratorControlPanelSchema(Interface):
     """
     """
@@ -56,11 +67,12 @@ class IDocumentGeneratorControlPanelSchema(Interface):
         default=safe_unicode(os.getenv('OO_SERVER', DEFAULT_OO_SERVER)),
     )
 
-    oo_port = schema.Int(
-        title=_(u'oo_port'),
-        description=_(u'Port Number of OO.'),
+    directives.widget("oo_port_list", DataGridFieldFactory, allow_reorder=True)
+    oo_port_list = schema.List(
+        title=_(u"oo_port_list"),
+        value_type=DictRow(title=u"oo_port_list", schema=IOOPortRowSchema),
         required=False,
-        default=int(os.getenv('OO_PORT', DEFAULT_OO_PORT))
+        default=[{"oo_port": int(os.getenv('OO_PORT', DEFAULT_OO_PORT))}]
     )
 
     uno_path = schema.TextLine(