+++ src/collective/documentgenerator/browser/generation_view.py
@@ -21,7 +21,7 @@ from collective.documentgenerator.content.pod_template import IPODTemplate
 from collective.documentgenerator.interfaces import (
     CyclicMergeTemplatesException, IDocumentFactory, PODTemplateNotFoundError)
 from collective.documentgenerator.utils import (remove_tmp_file,
-                                                temporary_file_name)
+                                                temporary_file_name, get_oo_port_list)
 
 from .. import _
 
@@ -261,7 +261,7 @@ class DocumentGenerationView(BrowserView):
             temp_filename,
             pythonWithUnoPath=config.get_uno_path(),
             ooServer=config.get_oo_server(),
-            ooPort=config.get_oo_port(),
+            ooPort=get_oo_port_list(),
             raiseOnError=raiseOnError,
             imageResolver=api.portal.get(),
             forceOoCall=True,