+++ src/collective/documentgenerator/config.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 from plone import api
 from Products.CMFPlone.utils import safe_unicode
 
@@ -57,12 +56,6 @@ def get_oo_server():
     )
 
 
-def get_oo_port():
-    return api.portal.get_registry_record(
-        'collective.documentgenerator.browser.controlpanel.IDocumentGeneratorControlPanelSchema.oo_port'
-    )
-
-
 def get_column_modifier():
     return api.portal.get_registry_record(
         'collective.documentgenerator.browser.controlpanel.IDocumentGeneratorControlPanelSchema.column_modifier'
@@ -106,14 +99,6 @@ def set_oo_server():
     if oo_server:
         api.portal.set_registry_record('collective.documentgenerator.browser.controlpanel.'
                                        'IDocumentGeneratorControlPanelSchema.oo_server', safe_unicode(oo_server))
-
-
-def set_oo_port():
-    """ Get environment value in buildout to define port """
-    oo_port = os.getenv('OO_PORT', DEFAULT_OO_PORT)
-    if oo_port:
-        api.portal.set_registry_record('collective.documentgenerator.browser.controlpanel.'
-                                       'IDocumentGeneratorControlPanelSchema.oo_port', int(oo_port))
 
 
 def set_uno_path():