+++ src/collective/documentgenerator/events/styles_events.py
@@ -14,7 +14,7 @@ from zope.i18n import translate
 from collective.documentgenerator import _, config
 from collective.documentgenerator.content.pod_template import IPODTemplate
 from collective.documentgenerator.utils import (remove_tmp_file,
-                                                temporary_file_name)
+                                                temporary_file_name, get_oo_port_list)
 
 logger = logging.getLogger('collective.documentgenerator: styles update')
 
@@ -86,7 +86,7 @@ def _update_template_styles(pod_template, style_template_filename):
                                                       tmp_file=temp_file.name,
                                                       extension='odt',
                                                       libreoffice_host=config.get_oo_server(),
-                                                      port=config.get_oo_port(),
+                                                      port=get_oo_port_list()[-1],  # the last one should be the less busy
                                                       style_template=style_template_filename,
                                                       stream=config.get_use_stream())
     (stdout, stderr) = executeCommand(cmd.split())